调整函数位置
